Remove events, overhaul Layers
Layer class layout is significantly simplified.  The basic layout is:
    class MyLayer(pyservice.Layer):
        def handle_request(self, context, next):
            ...
            next.handle_request(context)

            ...
This makes it much easier to reason about when each part of the layer
is invoked - previously, one would need to read the source for
handle_request to see when events were triggered.

Also cleaned up exception trapping a bit in the handler - more tests
needed.