Remove 'message' from exception output, add 'args'
Also putting the whole exception in a new object under the key
"__exception"

Got rid of the raise_ method of service because you should be able to
just
    raise MyException("string")
instead of
    service.raise_("MyException", "string")

Service.raise_ was only there to make sure that unknown exception
classes weren't made public.  Since that's in validate_exception
(and has been for some time) it serves no purpose.