Replace Stack class with execute function
The usual "class with __init__ and a function isn't a class" argument
applies here.  The only reason I preferred a class was for keeping
non-local variables (self.__index) when looping, but I can cheat that
into nested functions (and keep self!) by dumping the variable onto
the inner function and aliasing self to the inner function.

Good example here: http://stackoverflow.com/a/16032631