Simplify extension execution
The '__next' method on ExtensionExecutor is the magic bit here.  By
not requiring the items in the list to implement all operations, the
executor can just skip extensions that don't implement that step.
This is most immediately valuable to Service and Client, which want
to hook into handle_operation without doing anything for either of
before_operation or after_operation