Rewrite Client, rewrite config handling, simplify wsgi_application
Descriptions are gone.  Everything is loaded through the **api option
passed to Client/Service init.  Exception generation is cleaner -
the ClientProcessor will throw RequestException when the status code
is 4xx/5xx with the HTTP reason as the message.  NOTE: This is not
equivalent to wsgi.RequestException, this is a pure modeled exception,
similar to those specified in the provided api.