Large processor rename
Trying to unify some of the naming cruft that's been hanging around
since the very very early plugin and chaining designs.  For the most
part, names should now be one of: process*, enter/exit, execute, or
continue.

I also lost an hour or so trying to find a way to trim
frames from a traceback so that exceptions during processing didn't
have 80 nested "_continue() ... _continue() ... _execute()" to
scroll through.  Turns out it's not easy, and there are some
existing solutions that I'm not fond of.  There's a gist with
some more info on what I'd like to see:

https://gist.github.com/numberoverzero/89890832c1172b455084